minor refactor of credentials panel into its own file
