+++ pkg/gui/commit_message_panel.go
@@ -51,90 +51,6 @@ func (gui *Gui) handleCommitFocused(g *gocui.Gui, v *gocui.View) error {
 	return gui.renderString(g, "options", message)
 }
 
-type credentials chan string
-
-// waitForPassUname wait for a username or password input from the credentials popup
-func (gui *Gui) waitForPassUname(g *gocui.Gui, currentView *gocui.View, passOrUname string) string {
-	gui.credentials = make(chan string)
-	credentialsView, _ := g.View("credentials")
-	if passOrUname == "username" {
-		credentialsView.Title = gui.Tr.SLocalize("CredentialsUsername")
-		credentialsView.Mask = 0
-	} else {
-		credentialsView.Title = gui.Tr.SLocalize("CredentialsPassword")
-		credentialsView.Mask = '*'
-	}
-	g.Update(func(g *gocui.Gui) error {
-		err := gui.switchFocus(g, currentView, credentialsView)
-		if err != nil {
-			return err
-		}
-		gui.RenderCommitLength()
-		return nil
-	})
-
-	// wait for username/passwords input
-	userInput := <-gui.credentials
-	return userInput
-}
-
-func (gui *Gui) handlePushConfirm(g *gocui.Gui, v *gocui.View) error {
-	message := gui.trimmedContent(v)
-	if message == "" {
-		// make sure to input something
-		// if not dune the push progress will run forever
-		message = "-"
-	}
-	gui.credentials <- message
-	err := gui.refreshFiles(g)
-	if err != nil {
-		return err
-	}
-	v.Clear()
-	err = v.SetCursor(0, 0)
-	if err != nil {
-		return err
-	}
-	_, err = g.SetViewOnBottom("credentials")
-	if err != nil {
-		return err
-	}
-	nextView, err := gui.g.View("confirmation")
-	if err != nil {
-		nextView = gui.getFilesView(g)
-	}
-	err = gui.switchFocus(g, nil, nextView)
-	if err != nil {
-		return err
-	}
-	return gui.refreshCommits(g)
-}
-
-func (gui *Gui) handlePushClose(g *gocui.Gui, v *gocui.View) error {
-	_, err := g.SetViewOnBottom("credentials")
-	if err != nil {
-		return err
-	}
-
-	gui.credentials <- "-"
-	return gui.switchFocus(g, nil, gui.getFilesView(g))
-}
-
-func (gui *Gui) handleCredentialsViewFocused(g *gocui.Gui, v *gocui.View) error {
-	if _, err := g.SetViewOnTop("credentials"); err != nil {
-		return err
-	}
-
-	message := gui.Tr.TemplateLocalize(
-		"CloseConfirm",
-		Teml{
-			"keyBindClose":   "esc",
-			"keyBindConfirm": "enter",
-		},
-	)
-	return gui.renderString(g, "options", message)
-}
-
 func (gui *Gui) simpleEditor(v *gocui.View, key gocui.Key, ch rune, mod gocui.Modifier) {
 	switch {
 	case key == gocui.KeyBackspace || key == gocui.KeyBackspace2:
